Use Pratt parsing to solve operator precedence issue
Previously, precedence for the chain operator was too high
which would result in an expression like:
```
[!..!..!.] & [!...] => foo
```
being parsed as:
```
([!..!..!.] & ([!...] => foo))
```
which is rather unintuitive, I expect that most people would
expect the chain operator to have the lowest precedence and be
evaluated after the AND expression.

This commit introduces Pratt parsing (albeit very simplified)
to address this problem so now the order of operations is
more intuitive.

I've also inlined the postfix and infix functions into the
expression function to try and simplify the code a bit, expect
more cleanup changes to come.