Overhaul of MIDI sequencer
Big overhaul of the MIDI sequencer. Previously we would
loop over the timeline and push the events that needed
to be sent before waiting N ms to the next event. This
was actually fairly resource intensive (relatively speaking)
due to the implementation of the sleep which would wait
successively shorter times until we reached the target.

This was flawed for a number of reasons:
1. We needed to maintain 2 Event vectors which is redundant
2. The sequencer might be out of phase with the JACK callback
   causing increased latency
3. Some events nearer the end of the timeline could be dropped
   if JACK didn't manage to forward them to the port which often
   lead to "stuck notes" in your DAW/tool of choice

The new sequencer is implemented _inside_ the JACK callback which
means we can alleviate _all_ of these issues and even improve in
some regards like lower latency.