Use std::chrono for durations
Switched to using std::chrono for duration representation
instead of _assuming_ that time was measured in microseconds
everywhere. This should be a lot more typesafe not to mention
a lot more generic and should allow easy switching of timer
resolution.

All drift/inaccuracy should be eliminated now to a degree
that is reasonable and unnoticeable in most real-world
scenarios.