Layering is now the default behaviour for sinking
The layer operator has been removed in favour
of making its behaviour the default for the
sink operator.

So now the behaviour is that sinking sequences
to a channel will cause them to overlap instead
of concatenate. You can use the "wait" keyword
to cause them to be concatenated like:
```
wait !!...... @120 [52] ~> kick
wait ....!!.. @120 [52] ~> kick
```

This should make it much easier to use MIDI
drumkits which operate on the same channel
but use different notes.