Channel expressions
Added a third layer of expressions on top of
sequence expressions and literal expressions.

Channel expressions allow for operating
on the generated event timelines that
eventually get sent from the sequencer.

The implementation of these expressions
means we now have a lot of new operators
which make previously awkward problems
easy to solve and with an implementation
that is short and elegant.

I think with the introduction of channel
expressions, Cane is now ready to be used
for larger projects.

- Implemented channel expressions
- Removed notes (temporarily)
- Removed sync operator (to be replaced later with something better)
- Sequence variables now require $ prefix to disambiguate with channel expressions
- Definitions which allow for re-usable channel expressions
- A lot of cleanup and fixes
- Updated Kakoune highlighter
- Updated grammar