Immutable symbols & unique names
All symbols are now immutable (chains, definitions,
aliases and constants) and will error if you try
to re-define something.

All symbols must now be unique amongst every other
symbol kind. This means you cannot have an alias with
the same name as a constant for example.

Hopefully this will lead to clearer code.

Another notable change is that chains are now scoped
to channel expressions and will not be available to use
afterwards.