BPM is now a channel operator
Removed the sequence operator version of
BPM and replaced it with a channel expression
version. I think it makes more sense to set
the BPM when you're actually sending the sequence
out to the MIDI channel rather than attach it
to a sequence of steps.

Note mapping will be coming soon and operate
much the same way.