Major Overhaul
Reworked almost all of the compiler. Channel expressions have been
removed in favour of just using sequences. Channel expressions were
a little bit unintuitive to work with and added a lot of complexity
to the language for little gain plus the functionality they offered
overlapped significantly with existing features for sequence expressions.

- Update grammar
- Update Kakoune highlighter
- Remove hex and binary literals
- Remove +/- prefix operators for literal expressions
- Remove % operator for literal expressions
- Add global note value
- Remove channel expressions
- Add beats/skips for literal expressions to complement len
- Split implementation across multiple files for better organisation
- A lot of code cleanup
- Sequences now play in parallel and must be concatenated to play in sequence
- All includes have been removed from the implementation and instead are included in lib.hpp