Fix note mapping
Fixed an issue with note mapping that would cause it to map to _all_ notes,
even inactive notes which is undesirable. The documentation and code was
built to support the idea that note mapping should only map to active notes.
You can achieve the behaviour of mapping even to inactive notes by doing something
like this:
```
16:16 map note note+7 & 4:16
```