Step duration mapping with the BPM operator
The BPM operator (`@`) now maps duration to individual steps rather
than assinging a BPM to the entire sequence. This makes it more in line
with similiar operators `map` and `vel`.

Previously, you could not have differing BPM in the same sequence
which made things like nested tuplets pretty awkward to work with.
Now, you _can_ have differing BPMs in the same sequence which lets
you play around with a lot of interesting rhythms.

Also in this commit, the operator precedence of join (`,`) has
been lowered to the same level as the debug operator (`?`) along with
`map` and `vel`. The BPM operator (`@`) has been given its own precedence
level above `map`, `vel` and `,` which should make certain pieces of code
a bit more intuitive.
As an example:
```
4:16 @ 120, 6:16 @ 240 ~> foo
```
would have previously assigned the tempo `120` to the first sequence
and then later, the tempo `240` would be assigned to the entire sequence,
overriding the tempo of the first sequence. Now it gives each sequence
its own tempo and then joins them.