Remove channel aliases in favour of simply using 'let'
Remove channel aliases in favour of just having `let`. The two constructs
had a mostly overlapping usecases and the only real "advantage" of aliases
was that they disallowed arithmetic expressions but this is simply an
artificial limitation and it should be left to the user to decide what is
best. This simplifies the implmentation a bit and makes the language more
consistent.