Remove MIDI clock and global bpm/note
I was never happy with how the global bpm/note values worked. It felt
like a bit of a hack to allow MIDI clock to be calculated but the more
I thought about it, the more I realised that MIDI clock isn't really
a crucial feature to have for this kind of tool. It's useful for
sequencing hardware synths for sequences not intended for live-coding
but even then you could just use a sequencer with JACK support to
sequence your external hardware.

This change simplifies both the codebase and code written in Cane because
we no longer need to have a "metadata" section at the top of Cane source
files and no longer need to mess around with global BPM and notes. What
we do now instead is check that a sequence has note durations and notes
mapped when trying to send it to a MIDI channel. This actually helps
prevent bugs too because global note and bpm are no longer implicit.