Patterns
Add patterns which allow you to create re-usable bars and
cut down on duplication.

```
pat four_on_the_floor
	4:16 ~> kick $
	2:16 > 4 ~> snare $
	6:16 > 2 ~> hihat

\four_on_the_floor
\four_on_the_floor
```

The introduction of patterns required
a fairly substantial restructuring to the codebase for timelines.
The code is now a lot more modular and paves the way for more
features surrounding timelines.

- Update intro.md
- Update grammar to include patterns
- Update syntax highlighters to include patterns
- Add patterns